Share the cloud logo to the owner and status tables
